requirements.txt
python-dotenv
cloudinary
torch
diffusers
django
django-heroku
gunicorn